requirements.txt
Django==1.10.6
django-tinymce==2.6.0
olefile==0.44
Pillow==4.0.0
